% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Calculate the area under the curve of various plots including the receiver operating characteristic and precision-recall curve.}
\usage{
auc(preds, obs, curve = c("roc", "pr"), max_intervals = 1000)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{curve}{= c('roc', 'pr')}

\item{max_intervals}{The maximum number of thresholds to evaluate. Default = 1000.}
}
\value{
A named vector with the area under the curve of a specified plot.
}
\description{
Calculate the area under the curve of various plots including the receiver operating characteristic and precision-recall curve.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the AUC ROC
auc(predictions, observations, curve = 'roc')
}
