% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_diff.R
\name{boot_diff}
\alias{boot_diff}
\title{Calculate the bootstrapped empiric two-sided p-value and difference between two sets of predictions for a given performance metric.}
\usage{
boot_diff(
  preds1,
  preds2,
  obs,
  metric = NULL,
  reps = 1000,
  conf = 0.95,
  seed = NULL,
  ...
)
}
\arguments{
\item{preds1}{A vector of predicted probabilities for the first model.}

\item{preds2}{A vector of predicted probabilities for the second model.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{metric}{A function of the form f(preds, obs), e.g. `bs`}

\item{reps}{The number of bootstrap replicates. Default = 1000.}

\item{conf}{The width of the confidence interval. Default = 0.95.}

\item{seed}{An optional random seed.}

\item{...}{Additional arguments for the particular metric function, e.g. `thresh = 0.6`}
}
\description{
Calculate the bootstrapped empiric two-sided p-value and difference between two sets of predictions for a given performance metric.
}
\examples{
# Generate some predictions for two different models
p1 <- runif(1000)
p2 <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Confidence interval around the estimate of the Brier Score
boot_diff(p1, p2, observations, metric = brier)
}
