% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{brier}
\alias{brier}
\title{Calculate the Brier Score for predicted probabilities against a binary outcome.}
\usage{
brier(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The Brier Score given by \deqn{ BS = \frac{1}{N} \sum (y - \hat y)^2}
}
\description{
Calculate the Brier Score for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Brier Score
brier(predictions, observations)
}
