% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.ci.R
\name{bs.ci}
\alias{bs.ci}
\title{Calculate the bias-corrected, basic bootstrapped confidence interval for a given performance metric.}
\usage{
bs.ci(preds, obs, metric = NULL, reps = 1000, conf = 0.95,
  seed = NULL)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{metric}{A function of the form f(preds, obs), e.g. `bs`}

\item{reps}{The number of bootstrap replicates. Default = 1000.}

\item{conf}{The width of the confidence interval. Default = 0.95.}

\item{seed}{An optional random seed.
#' @examples
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Confidence interval around the estimate of the Brier Score
bs.ci(predictions, observations, bs)}
}
\description{
Calculate the bias-corrected, basic bootstrapped confidence interval for a given performance metric.
}
