% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvd.R
\name{bvd}
\alias{bvd}
\title{Decompose the prediction error into bias and variance components}
\usage{
bvd(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
`bias` given by  \deqn{\frac{1}{N} \sum (\frac{1}{N} \sum(\hat y)) - y)^2}

`variance` given by  \deqn{\frac{1}{N} \sum (\frac{1}{N} \sum(\hat y)) - \hat y)^2}
}
\description{
Decompose the prediction error into bias and variance components
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the bias and variance
bvd(predictions, observations)
}
