% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_plot.R
\name{calib_plot}
\alias{calib_plot}
\title{Produce a calibration plot for a set of predixcted probabilities for a binary classifer.}
\usage{
calib_plot(form, data, cuts = 10, refline = TRUE, smooth = FALSE,
  rug = FALSE)
}
\arguments{
\item{form}{A formula where the left-hand side is the variable representing the observed outcome, 0 or 1, and the right-hand side represents the column names of the different model probabilities.}

\item{data}{A data frame that contains at least two columns, one of which is the observed outcome and the others that are predicted probabilities.}

\item{cuts}{The number of bins of probabilities. Default = 10.}

\item{refline}{Whether or not to include a 45 degree reference line. Defailt = TRUE.}

\item{smooth}{Whether or not to include a smoothed line for each models' probabilities. Default = FALSE.}

\item{rug}{Whether or not to include a rug plot of the observed probabilities. Usually works best with only one model. Defailt = FALSE.}
}
\description{
Produce a calibration plot for a set of predixcted probabilities for a binary classifer.
}
\examples{
m1 <- glm(mpg > 20 ~ cyl + disp + hp, family = 'binomial', data = mtcars)
results <- data.frame(outcome = mtcars$mpg > 20, lr_1 = predict(m1, type = 'response'))
calib_plot(outcome ~ lr_1, data = results)
}
