% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstat.R
\name{cstat}
\alias{cstat}
\title{Calculate the C-statistic for predicted probabilities against a binary outcome.}
\usage{
cstat(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The C-statistic as a single numeric value. This value is equivalent to the area under the curve of the receiver operatoing characteristic.
}
\description{
Calculate the C-statistic for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the C-statistic
cstat(predictions, observations)
}
