% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscore.R
\name{fscore}
\alias{fscore}
\title{Calculate the F Score for predicted probabilities against a binary outcome.}
\usage{
fscore(preds, obs, beta = 1, thresh = 0.5)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{beta}{The default weighting of the precision and recall. Default beta = 1 which yields the "F_1" Score.}

\item{thresh}{The probability threshold at or above which a prediction is considered to be positive.}
}
\value{
The F Score is the harmonic mean of the precision (positive predictive value) and recall (sensitivity).
}
\description{
Calculate the F Score for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Scaled Brier Score
fscore(predictions, observations)
}
