% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ici.r
\name{ici}
\alias{ici}
\title{Calculate the integrated calibration index (ICI) for predicted probabilities against a binary outcome. Based on Austin PC, Steyerberg EW. The Integrated Calibration Index (ICI) and related metrics for quantifying the calibration of logistic regression models. Statistics in Medicine. 2019;1–15. https://doi.org/10.1002/sim.8281}
\usage{
ici(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The integrated calibration index (ICI)
}
\description{
Calculate the integrated calibration index (ICI) for predicted probabilities against a binary outcome. Based on Austin PC, Steyerberg EW. The Integrated Calibration Index (ICI) and related metrics for quantifying the calibration of logistic regression models. Statistics in Medicine. 2019;1–15. https://doi.org/10.1002/sim.8281
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Integrated Calibration Index
ici(predictions, observations)
}
