% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lloss.R
\name{lloss}
\alias{lloss}
\title{Calculate the Log Lossfor predicted probabilities against a binary outcome.}
\usage{
lloss(preds, obs, eps = 1e-15)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The Log Loss given by \deqn{ logloss = y_i(\log \hat y_i) + (1-y_i)\log(1-\hat y_i)}
}
\description{
Calculate the Log Lossfor predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Brier Score
lloss(predictions, observations)
}
