% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_perf_df.R
\name{make_perf_df}
\alias{make_perf_df}
\title{Take a list of performance metrics and return the observed estimate and 95\% confidence intervals based on 1000 bootstrap replicates in the form a data.frame.}
\usage{
make_perf_df(preds, obs, metrics = list(brier, sbrier, ici, lloss, cstat), ...)
}
\arguments{
\item{preds}{A vector of predicted probabilities for the first model.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{metrics}{A list of function of the form f(preds, obs), e.g. `bs`}

\item{...}{Additional arguments for the particular metric or boot function, e.g. `thresh = 0.6`}
}
\description{
Take a list of performance metrics and return the observed estimate and 95\% confidence intervals based on 1000 bootstrap replicates in the form a data.frame.
}
\examples{
# Generate some predictions for two different models
preds <- runif(1000)
# Generate some binary outcomes
obs <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Confidence interval around the estimate of the Brier Score
make_perf_df(preds, obs, metrics = list(brier, cstat))
}
