% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sbrier.R
\name{mc_sbrier}
\alias{mc_sbrier}
\title{Calculate the multiclass scaled Brier Score for predicted probabilities against a binary outcome.}
\usage{
mc_sbrier(preds, obs)
}
\arguments{
\item{preds}{A data.frame or matrix of predicted probabilities with one column per class.}

\item{obs}{A data.frame or matrix containing the observed binary outcomes (0 or 1) with one column per class. The order of the columns should match the order for preds.}
}
\value{
The multiclass scaled Brier Score given by \deqn{1 - \frac{BS}{BS_{max}}}, where the multiclass Brier score is given by \deqn{ BS = \frac{1}{N} \sum^K \sum^N (y - \hat y)^2} for N observations and K classes
}
\description{
Calculate the multiclass scaled Brier Score for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- data.frame(p1 = runif(1000), p2 = runif(1000), p3 = runif(1000))
# Generate some binary outcomes
observations <- data.frame(o1 = sample(0:1, size = 1000, replace = TRUE), o2 = sample(0:1, size = 1000, replace = TRUE), o3 = sample(0:1, size = 1000, replace = TRUE))
# Calculate the multiclass Brier Score
mc_sbrier(predictions, observations)
}
