% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_avg.R
\name{micro_avg}
\alias{micro_avg}
\title{Calculate the micro-averaged performance of some measure against a binary outcome across a range of classes in a multi-label classification task.}
\usage{
micro_avg(preds, obs, metric = cstat, ...)
}
\arguments{
\item{preds}{A matrix of predicted probabilities for each class.}

\item{obs}{A matrix containing the observed binary outcomes (0 or 1) for each class. Columns should occur in the same order for `preds` and `obs`.}

\item{metric}{A function name for the performance metric of interest. Default is `cstat`.}

\item{...}{Additional parameters that can be passed to `metric` as needed.}
}
\value{
The micro-averaged `metric` as a single numeric value.
}
\description{
Calculate the micro-averaged performance of some measure against a binary outcome across a range of classes in a multi-label classification task.
}
\examples{
# Generate some predictions
predictionsA <- runif(1000)
predictionsB <- runif(1000)
# Generate some binary outcomes
observationsA <- sample(0:1, size = 1000, replace = TRUE)
observationsB <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the C-statistic
micro_avg(matrix(c(predictionsA, predictionsB), ncol = 2),
          matrix(c(observationsA, observationsB), ncol = 2))
}
