% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Calculate the mean-squared error for predicted vs observed continuous outcomes.}
\usage{
mse(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted continuous outcomes.}

\item{obs}{A vector containing the observed continuous outcomes.}
}
\value{
The Mean Squared Error (MSE) given by \deqn{ MSE = \frac{1}{N} \sum (y - \hat y)^2}
}
\description{
Calculate the mean-squared error for predicted vs observed continuous outcomes.
}
\examples{
# Generate some predictions
predictions <- rnorm(1000)
# Generate some observed outcomes
observations <- rnorm(1000)
# Calculate the MSE
mse(predictions, observations)
}
