% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb}
\alias{nb}
\title{Calculate the net benefit of a model at a given probability threshold.}
\usage{
nb(preds, obs, p_t, weight = NULL)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{p_t}{The probability threshold at or above which a prediction is considered to be positive.}

\item{weight}{Relative weighted importance of true positives to false positives independent of the classification threshold. When weight = NULL, the original net benefit calculation is used. Default = NULL}
}
\value{
The true positive count is the number of observations predicted as positive that are indeed positive.
}
\description{
Calculate the net benefit of a model at a given probability threshold.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the true positive count
nb(predictions, observations, p_t = 0.25)
}
