% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocp.R
\name{ocp}
\alias{ocp}
\title{Calculate the optimism-corrected performance of a model using bootstrap validation. Approach from Steyerberg EW. Clinical Prediction Models, 2nd Edition. pp 106--107}
\usage{
ocp(mwf, form, data, metric = sbrier, reps = 250, ...)
}
\arguments{
\item{mwf}{A function that represents the entire modeling workflow. Should take arguments ``form` and `data` fit a model then return it to make predictions.}

\item{form}{A formula describing the model.}

\item{data}{A data frame including all variables in `form`.}

\item{reps}{The number of bootstrap replicates. Default = 250.}

\item{...}{Additional arguments to make correct predictions, for example, `type = response` when needed}

\item{metrics}{A performance metric as function that takes (`preds`, `obs`) to calculate. Default = `sbrier`}
}
\description{
Calculate the optimism-corrected performance of a model using bootstrap validation. Approach from Steyerberg EW. Clinical Prediction Models, 2nd Edition. pp 106--107
}
\examples{
glm_mwf <- \(form, data) { glm(form, data = data, family = binomial) }
mymtcars <- mtcars
mymtcars$mpg20 <- mymtcars$mpg > 20
ocp(glm_mwf, form = mpg20 ~ cyl + disp + hp, data = mymtcars, type = 'response')

}
