% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe.df.R
\name{ohe.df}
\alias{ohe.df}
\title{One-hot encodes all categorical variables in a data.frame}
\usage{
ohe.df(df, drop_ref = TRUE)
}
\arguments{
\item{df}{A data.frame that can contain columns of any variable type.}

\item{drop_ref}{Logical, default = \code{TRUE}. Will return x - 1 columns for x unique categories. Choose \code{FALSE} to return x columns.}
}
\value{
A data.frame of all columns of class factor returned as one-hot encoded. All other column types are returned untouched.
}
\description{
One-hot encodes all categorical variables in a data.frame
}
\examples{
df <- data.frame(lets = rep(letters[1:3], 3), LETS = rep(LETTERS[1:3], 3), x = rnorm(9))
ohe.df(df)
}
