% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe.R
\name{ohe}
\alias{ohe}
\title{One-hot encodes a single data.frame column or vector}
\usage{
ohe(var, drop_ref = TRUE)
}
\arguments{
\item{var}{A vector or data.frame column of categorical variables}

\item{drop_ref}{Logical, default = \code{TRUE}. Will return x - 1 columns for x unique categories. Choose \code{FALSE} to return x columns.}
}
\value{
A data.frame of columns each representing a category in \code{var} with binary indicators 0/1.
}
\description{
One-hot encodes a single data.frame column or vector
}
\examples{
# Encode a vector
lets <- rep(letters[1:3], 3)
ohe(lets)

# Encode a single column of a data.frame
df <- data.frame(lets, x = runif(9))
ohe(df$lets)

# Encode just the categorical columns and combine
df2 <- data.frame(lets, LETS = rep(LETTERS[1:3], 3), x = rnorm(9))
data.frame(lapply(df2, function(x) if(is.factor(x)) { ohe(x) } else { x } ))
}
