% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_plot.R
\name{pr_plot}
\alias{pr_plot}
\title{Produce a precision-recall plot for a set of predicted probabilities for a binary classifier.}
\usage{
pr_plot(form, data, smooth = FALSE)
}
\arguments{
\item{form}{A formula where the left-hand side is the variable representing the observed outcome, 0 or 1, and the right-hand side represents the column names of the different model probabilities.}

\item{data}{A data frame that contains at least two columns, one of which is the observed outcome and the others that are predicted probabilities.}

\item{smooth}{Whether or not to include a smoothed line for each models' probabilities. Default = FALSE.}
}
\description{
Produce a precision-recall plot for a set of predicted probabilities for a binary classifier.
}
\examples{
library(ranger)
library(palmerpenguins)
pp <- penguins[complete.cases(penguins),]
m1 <- ranger(species == 'Adelie' ~ island + bill_length_mm + flipper_length_mm + body_mass_g + sex,
      data = pp, probability = TRUE)
p_obj <- predict(m1, data = pp)
results <- data.frame(preds_m1 = p_obj$predictions[,2],
                       obs = pp$species == 'Adelie')
pr_plot(obs ~ preds_m1, data = results)
}
