% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq}
\alias{rsq}
\title{Calculate the R^2 of a model for predicted probabilities against a binary outcome.}
\usage{
rsq(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The R^2 value
}
\description{
Calculate the R^2 of a model for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the R^2
rsq(predictions, observations)
}
