% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbrier.R
\name{sbrier}
\alias{sbrier}
\title{Calculate the Scaled Brier Score for predicted probabilities against a binary outcome.}
\usage{
sbrier(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\value{
The Scaled Brier Score given by \deqn{ BS_{scaled} = 1 - \frac{BS}{BS_{max}}} where
\deqn{BS_{max} = \frac{1}{N} \sum (y - \bar y)^2}.
}
\description{
Calculate the Scaled Brier Score for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Scaled Brier Score
sbrier(predictions, observations)
}
