% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens}
\alias{sens}
\title{Calculate the Sensitivity for predicted probabilities against a binary outcome.}
\usage{
sens(preds, obs, thresh = 0.5)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{thresh}{The threshold at or above which a prediction is considered positive. Default = 0.5.}
}
\value{
The sensitivity is the probability of a true prediction conditional on a true outcome.
}
\description{
Calculate the Sensitivity for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Sensitivity
sens(predictions, observations)
# And at a different threshold
sens(predictions, observations, thresh = 0.8)
}
