% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_NNE.R
\name{thresh_NNE}
\alias{thresh_NNE}
\title{Index observation(s) in a sorted set of prediction values for a set of NNE thresholds.}
\usage{
thresh_NNE(threshold, PPV)
}
\arguments{
\item{threshold}{A set of NNE thresholds.}

\item{PPV}{A vector of precision values (NNE=1/PPV) at a set of yhat values sorted in descending order.}
}
\description{
Index observation(s) in a sorted set of prediction values for a set of NNE thresholds.
}
\examples{
# Generate some predictions
PPV <- runif(1000)
# Order the PPV sample
o <- order(PPV, decreasing=T)
# Identify yhat values
thresh_NNE(thresh=seq(5,15,1), PPV=PPV[o])
}
