% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_plot.R
\name{thresh_plot}
\alias{thresh_plot}
\title{Produce a precision-recall plot for a set of predicted probabilities for a binary classifier.}
\usage{
thresh_plot(
  preds,
  obs,
  metrics = c("ppv", "npv", "fscore"),
  max_intervals = 1000
)
}
\arguments{
\item{max_intervals}{The maximum number of thresholds to evaluate. Default = 1000.}

\item{form}{A formula where the left-hand side is the variable representing the observed outcome, 0 or 1, and the right-hand side represents the column names of the different model probabilities.}

\item{data}{A data frame that contains at least two columns, one of which is the observed outcome and the others that are predicted probabilities.}
}
\description{
Produce a precision-recall plot for a set of predicted probabilities for a binary classifier.
}
\examples{
library(ranger)
library(palmerpenguins)
pp <- penguins[complete.cases(penguins),]
m1 <- ranger(species == 'Adelie' ~ island + bill_length_mm + flipper_length_mm + body_mass_g + sex,
      data = pp, probability = TRUE)
p_obj <- predict(m1, data = pp)
results <- data.frame(preds_m1 = p_obj$predictions[,2],
                       obs = pp$species == 'Adelie')
thresh_plot(p_obj$predictions[,2], pp$species == 'Adelie')
}
