% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_PPV.R
\name{thresh_PPV}
\alias{thresh_PPV}
\title{Index observation(s) in a sorted set of prediction values for a set of PPV thresholds.}
\usage{
thresh_PPV(threshold, PPV)
}
\arguments{
\item{threshold}{A set of NNE thresholds.}

\item{PPV}{A vector of precision values at a set of yhat values sorted in descending order.}
}
\description{
Index observation(s) in a sorted set of prediction values for a set of PPV thresholds.
}
\examples{
# Generate some predictions
PPV <- runif(1000)
# Order the PPV sample
o <- order(PPV, decreasing=T)
# Identify yhat values
thresh_PPV(thresh=seq(5,15,1), PPV=PPV[o])
}
