% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_sens.R
\name{thresh_sens}
\alias{thresh_sens}
\title{Index observation(s) in a sorted set of prediction values for a set of sensitivity thresholds.}
\usage{
thresh_sens(threshold, sensitivity)
}
\arguments{
\item{threshold}{A set of NNE thresholds.}

\item{sensitivity}{A vector of precision values at a set of yhat values sorted in descending order.}
}
\description{
Index observation(s) in a sorted set of prediction values for a set of sensitivity thresholds.
}
\examples{
# Generate some predictions
sensitivity <- runif(1000)
# Order the sensitivity sample
o <- order(sensitivity, decreasing=T)
# Identify yhat values
thresh_sens(thresh=seq(5,15,1), PPV=PPV[o])
}
