% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_yhat.R
\name{thresh_yhat}
\alias{thresh_yhat}
\title{Index observation(s) in a sorted set of prediction values for a set of prediction threshold.}
\usage{
thresh_yhat(threshold, yhat)
}
\arguments{
\item{threshold}{A set of prediction thresholds.}

\item{yhat}{A vector of predicted probabilities.}
}
\description{
Index observation(s) in a sorted set of prediction values for a set of prediction threshold.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Order the predictions
o <- order(predictions, decreasing=T)
# Identify yhat values
thresh_yhat(thresh=seq(0.2,0.4,0.05), yhat=predictions[o])
}
