% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshstats.R
\name{threshstats}
\alias{threshstats}
\title{Calculate Threshold Metrics in a sorted set of prediction values at every observation.}
\usage{
threshstats(preds, obs)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}
}
\description{
Calculate Threshold Metrics in a sorted set of prediction values at every observation.
}
\examples{
# Generate some predictions
preds <- runif(1000)
# Generate some binary outcomes
obs <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the false negative count
threshstats(preds, obs)
}
