% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zol.R
\name{zol}
\alias{zol}
\title{Calculate the zero-one loss (i.e. accuracy) for predicted probabilities against a binary outcome.}
\usage{
zol(preds, obs, thresh = 0.5)
}
\arguments{
\item{preds}{A vector of predicted probabilities.}

\item{obs}{A vector containing the observed binary outcomes (0 or 1).}

\item{thresh}{The threshold at or above which a prediction is considered positive. Default = 0.5.}
}
\value{
The accuracy, or zero-one loss, of the predictions at a given threshold.
}
\description{
Calculate the zero-one loss (i.e. accuracy) for predicted probabilities against a binary outcome.
}
\examples{
# Generate some predictions
predictions <- runif(1000)
# Generate some binary outcomes
observations <- sample(0:1, size = 1000, replace = TRUE)
# Calculate the Brier Score
zol(predictions, observations)
}
